Avtivates the tests
half of them fails at the moment, probably because I don't return
anything since mapping is not a get one

+++ src/main/java/com/workintech/zoo/entity/Koala.java
@@ -11,7 +11,7 @@ public class Koala {
     private long id;
     private String name;
     private double weight;
-    private int sleepHour;
+    private double sleepHour;
     private String gender;
